+++ src/net/craftlin/bukkit/impl/entity/base/BukkitLivingEntity.kt
@@ -28,12 +28,12 @@ abstract class BukkitLivingEntity(private val livingEntity: org.bukkit.entity.Li
 
     override val effects: Collection<Effect>
         get() = livingEntity.activePotionEffects.map {
-            Effect(BukkitEffectType.Converter(it.type), it.duration.toLong(), it.amplifier.toLong(), it.isAmbient, it.hasParticles(), it.hasIcon())
+            Effect(BukkitEffectType.Converter(it.type), it.duration.toLong(), it.amplifier, it.isAmbient, it.hasParticles(), it.hasIcon())
         }
 
     override fun effect(effect: Effect) {
         livingEntity.addPotionEffect(with(effect) {
-            PotionEffect(BukkitEffectType.Converter(type), time.toInt(), strength.toInt(), ambient, particles, icon)
+            PotionEffect(BukkitEffectType.Converter(type), time.toInt(), strength, ambient, particles, icon)
         })
     }
